+++ packages/react/src/components/Web3Button/index.tsx
@@ -1,8 +1,8 @@
-import { useActiveChainId } from "../../Provider";
 import { useContract } from "../../hooks/async/contracts";
-import { useAddress } from "../../hooks/useAddress";
-import { useChainId } from "../../hooks/useChainId";
-import { useNetwork } from "../../hooks/useNetwork";
+import { useNetworkMismatch } from "../../hooks/useNetworkMismatch";
+import { useNetwork } from "../../hooks/wagmi-required/useNetwork";
+import { useAddress, useChainId } from "../../hooks/wallet";
+import { useSDKChainId } from "../../providers/base";
 import {
   createCacheKeyWithNetwork,
   createContractCacheKey,
@@ -10,6 +10,7 @@ import {
 import { ConnectWallet } from "../ConnectWallet";
 import { Button } from "../shared/Button";
 import { ThemeProvider, ThemeProviderProps } from "../shared/ThemeProvider";
+import { FiWifi } from "@react-icons/all-files/fi/FiWifi";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import type { SmartContract } from "@thirdweb-dev/sdk/dist/declarations/src/contracts/smart-contract";
 import type { CallOverrides } from "ethers";
@@ -20,6 +21,7 @@ type ActionFn = (contract: SmartContract) => any;
 
 interface Web3ButtonProps<TActionFn extends ActionFn>
   extends ThemeProviderProps {
+  className?: string;
   contractAddress: `0x${string}` | `${string}.eth` | string;
 
   overrides?: CallOverrides;
@@ -64,14 +66,17 @@ export const Web3Button = <TAction extends ActionFn>({
   isDisabled,
   children,
   action,
+  className,
   ...themeProps
 }: PropsWithChildren<Web3ButtonProps<TAction>>) => {
   const address = useAddress();
   const walletChainId = useChainId();
-  const sdkChainId = useActiveChainId();
+  const sdkChainId = useSDKChainId();
   const [, switchNetwork] = useNetwork();
 
   const queryClient = useQueryClient();
+
+  const hasMismatch = useNetworkMismatch();
 
   const switchToChainId = useMemo(() => {
     if (sdkChainId && walletChainId && sdkChainId !== walletChainId) {
@@ -87,7 +92,7 @@ export const Web3Button = <TAction extends ActionFn>({
       if (switchToChainId) {
         if (switchNetwork) {
           await switchNetwork(switchToChainId);
-          await new Promise((resolve) => setTimeout(resolve, 500));
+          return "__NETWORK_SWITCHED__";
         } else {
           throw new Error(
             "need to switch chain but connected wallet does not support switching",
@@ -99,10 +104,14 @@ export const Web3Button = <TAction extends ActionFn>({
       if (onSubmit) {
         onSubmit();
       }
+
       return await action(contract);
     },
     {
       onSuccess: (res) => {
+        if (res === "__NETWORK_SWITCHED__") {
+          return;
+        }
         if (onSuccess) {
           onSuccess(res);
         }
@@ -122,18 +131,24 @@ export const Web3Button = <TAction extends ActionFn>({
     },
   );
   if (!address) {
-    return <ConnectWallet {...themeProps} />;
+    return <ConnectWallet className={className} {...themeProps} />;
   }
+
+  const willSwitchNetwork = hasMismatch && !!switchNetwork;
 
   return (
     <ThemeProvider {...themeProps}>
       <Button
-        style={{ height: "50px" }}
+        className={className}
+        style={{ height: "50px", minWidth: "200px", width: "100%" }}
         isLoading={mutation.isLoading || !contract}
         onClick={() => mutation.mutate()}
-        isDisabled={isDisabled}
+        isDisabled={willSwitchNetwork ? false : isDisabled}
+        leftElement={
+          willSwitchNetwork ? <FiWifi width="1em" height="1em" /> : undefined
+        }
       >
-        {children}
+        {willSwitchNetwork ? "Switch Network" : children}
       </Button>
     </ThemeProvider>
   );