+++ packages/react/src/solana/hooks/useSDK.tsx
@@ -1,0 +1,6 @@
+import { ThirdwebSDKContext } from "../providers/ThirdwebSDKProvider";
+import { useContext } from "react";
+
+export function useSDK() {
+  return useContext(ThirdwebSDKContext);
+}