+++ packages/sdk/src/schema/contracts/custom.ts
@@ -1,5 +1,10 @@
 import { toSemver } from "../../common/index";
-import { AddressSchema, BigNumberishSchema } from "../shared";
+import {
+  AddressSchema,
+  BigNumberishSchema,
+  BigNumberTransformSchema,
+  FileOrBufferOrStringSchema,
+} from "../shared";
 import {
   CommonContractOutputSchema,
   CommonContractSchema,
@@ -10,7 +15,6 @@ import {
   CommonTrustedForwarderSchema,
   MerkleSchema,
 } from "./common";
-import { FileOrBufferOrStringSchema, JsonSchema } from "@thirdweb-dev/storage";
 import { BigNumberish } from "ethers";
 import { z } from "zod";
 
@@ -18,7 +22,7 @@ import { z } from "zod";
  * @internal
  */
 export const BYOCContractMetadataSchema = CommonContractSchema.catchall(
-  z.lazy(() => JsonSchema),
+  z.union([BigNumberTransformSchema, z.unknown()]),
 );
 
 /**