+++ packages/sdk/src/schema/tokens/common/index.ts
@@ -1,6 +1,10 @@
-import { BigNumberSchema, HexColor } from "../../shared";
+import {
+  BigNumberSchema,
+  BigNumberTransformSchema,
+  FileOrBufferOrStringSchema,
+  HexColor,
+} from "../../shared";
 import { OptionalPropertiesInput } from "./properties";
-import { FileOrBufferOrStringSchema, JsonSchema } from "@thirdweb-dev/storage";
 import { z } from "zod";
 
 /**
@@ -13,7 +17,7 @@ export const CommonTokenInput = z
     image: FileOrBufferOrStringSchema.nullable().optional(),
     external_url: FileOrBufferOrStringSchema.nullable().optional(),
   })
-  .catchall(z.lazy(() => JsonSchema));
+  .catchall(z.union([BigNumberTransformSchema, z.unknown()]));
 
 /**
  * @internal