+++ packages/solana/src/types/common.ts
@@ -1,5 +1,4 @@
 import { Signer, WalletAdapter } from "@metaplex-foundation/js";
-import { JsonObjectSchema } from "@thirdweb-dev/storage";
 import { z } from "zod";
 
 /**
@@ -36,6 +35,21 @@ export const JsonLiteral = z.union([
 /**
  * @internal
  */
+export type JsonLiteral = number | string | null | boolean;
+
+/**
+ * @internal
+ */
+export type Json = JsonLiteral | JsonObject | Json[];
+
+/**
+ * @internal
+ */
+export type JsonObject = { [key: string]: Json };
+
+/**
+ * @internal
+ */
 export const HexColor = z.union([
   z
     .string()
@@ -47,8 +61,13 @@ export const HexColor = z.union([
 /**
  * @internal
  */
+const PropertiesInput = z.object({}).catchall(z.unknown());
+
+/**
+ * @internal
+ */
 export const OptionalPropertiesInput = z
-  .union([z.array(JsonObjectSchema), JsonObjectSchema])
+  .union([z.array(PropertiesInput), PropertiesInput])
   .optional();
 
 /**
@@ -104,3 +123,27 @@ export type WalletAccount = {
   address: string;
   name: string;
 };
+
+const isBrowser = () => typeof window !== "undefined";
+const FileOrBufferUnionSchema = isBrowser()
+  ? (z.instanceof(File) as z.ZodType<InstanceType<typeof File>>)
+  : (z.instanceof(Buffer) as z.ZodTypeAny); // @fixme, this is a hack to make browser happy for now
+
+/**
+ * @internal
+ */
+export const FileOrBufferSchema = z.union([
+  FileOrBufferUnionSchema,
+  z.object({
+    data: z.union([FileOrBufferUnionSchema, z.string()]),
+    name: z.string(),
+  }),
+]);
+
+/**
+ * @internal
+ */
+export const FileOrBufferOrStringSchema = z.union([
+  FileOrBufferSchema,
+  z.string(),
+]);